Moved code to nav loading fragment and pass nav to the navigation page.
