+++ app/src/main/java/edu/cascadia/mobas/campusguidebook/ui/home/HomeFragment.java
@@ -35,11 +35,6 @@ public class HomeFragment extends Fragment {
     protected MainActivityViewModel getViewModel() {
         return mViewModel;
     }
-    private List<Location> mLocations;
-
-
-
-
 
     // All UI behavior has moved from MainActivity to fragments
     // View initialization logic goes in onCreateView
@@ -51,28 +46,7 @@ public class HomeFragment extends Fragment {
         binding = HomeBinding.inflate(inflater, container, false);
         mViewModel = new ViewModelProvider(this.requireActivity()).get(MainActivityViewModel.class);
 
-        new AsyncTask<Void, Void, List<Location>>() {
-            @Override
-            protected List<Location> doInBackground(Void... voids) {
-                mLocations = mViewModel.getAllLocations();
-                return mLocations;
-            }
 
-            @SuppressLint("StaticFieldLeak")
-            @Override
-            protected void onPostExecute(List<Location> locations) {
-                //TODO: For Testing:
-
-                //List<String> path = calculate("Parking 1","15 Min Parking 1");
-                List<String> path = calculate("Parking 1","Bus stop 1");
-                //List<String> path = calculate("Parking 1","CC01 Entry LL");
-                Integer i = 0;
-                for (String item : path) {
-                    i++;
-                    Log.d("Nav", "0xB45 - Nav step " + i + ":  " + item);
-                }
-            }
-        }.execute();
         return binding.getRoot();
     }
 
@@ -90,12 +64,14 @@ public class HomeFragment extends Fragment {
                     .navigate(R.id.nav_sust_list);
         });
         binding.cardViewHomeClubs.setOnClickListener(CardView -> {
-                Navigation.findNavController(binding.cardViewHomeClubs)
-                        .navigate(R.id.action_nav_home_to_nav_clubs);
+            Navigation.findNavController(binding.cardViewHomeClubs)
+                    .navigate(R.id.action_nav_home_to_nav_clubs);
+        });
+        binding.cardViewHomeUNUSED4.setOnClickListener(CardView -> {
+            Navigation.findNavController(binding.cardViewHomeUNUSED4)
+                    .navigate(R.id.action_nav_home_to_to_navigation_loading);
         });
 
-
-        
     }
 
     @Override
@@ -103,124 +79,5 @@ public class HomeFragment extends Fragment {
         super.onDestroyView();
         binding = null;
     }
-    List<String> Steps = new ArrayList<>();
-    List<Integer> excludedNumbers = new ArrayList<>();
-    Location currentLocation;
-    Integer exeCount = 0;
-
-    boolean returnedPath = false;
-    public List<String> calculate (String mStart, String mEnd){
-
-        List<String> mPath = calculatePath(mStart, mEnd);
-
-        while (mPath == null && returnedPath == false) {
-            Log.d("t", "0x84WQ - Restarting Path");
-            Steps.clear();
-            excludedNumbers.clear();
-            currentLocation = null;
-            mPath = calculatePath(mStart, mEnd);
-        }
-
-        returnedPath = true;
-        return mPath;
-    }
-    public List<String> calculatePath (String Start, String End) {
-        currentLocation = getLocation(Start);
-
-
-        List<String> directions = Arrays.asList(
-                currentLocation.getNorth(),
-                currentLocation.getNorthEast(),
-                currentLocation.getEast(),
-                currentLocation.getSouthEast(),
-                currentLocation.getSouth(),
-                currentLocation.getSouthWest(),
-                currentLocation.getWest(),
-                currentLocation.getNorthWest()
-        );
-
-        while (directions.stream().noneMatch(direction -> direction.equals(End))) {
-            //Log.d("t", "0x89D3" + Steps);
-            directions = Arrays.asList(
-                    currentLocation.getNorth(),
-                    currentLocation.getNorthEast(),
-                    currentLocation.getEast(),
-                    currentLocation.getSouthEast(),
-                    currentLocation.getSouth(),
-                    currentLocation.getSouthWest(),
-                    currentLocation.getWest(),
-                    currentLocation.getNorthWest()
-            );
-            //if (Steps.size() == 2) {
-            //Log.d("t", "0x89DA   " + currentLocation.getName() + "  Dir:  " + directions);
-            //}
-            switch (getRand(excludedNumbers)) {
-                case 1:
-                    moveAndAddToHistory(currentLocation.getNorth());
-                    break;
-                case 2:
-                    moveAndAddToHistory(currentLocation.getNorthWest());
-                    break;
-                case 3:
-                    moveAndAddToHistory(currentLocation.getNorthEast());
-                    break;
-                case 4:
-                    moveAndAddToHistory(currentLocation.getEast());
-                    break;
-                case 5:
-                    moveAndAddToHistory(currentLocation.getSouthEast());
-                    break;
-                case 6:
-                    moveAndAddToHistory(currentLocation.getSouth());
-                    break;
-                case 7:
-                    moveAndAddToHistory(currentLocation.getSouthWest());
-                    break;
-                case 8:
-                    moveAndAddToHistory(currentLocation.getWest());
-                    break;
-                default:
-            }
-            exeCount++;
-            if (exeCount > 30){
-                exeCount = 0;
-                return null;
-            }
-
-        }
-        if (!directions.stream().noneMatch(direction -> direction.equals(End))){
-            Steps.add(currentLocation.getName());
-        }
-
-        Steps.add(End);
-        return Steps;
-    }
-
-
-    public Integer getRand (List<Integer> excludedNumbers){
-        Random random = new Random();
-        int randomNumber;
-        do {
-            randomNumber = random.nextInt(8) + 1;
-        } while (excludedNumbers.contains(randomNumber));
-        return randomNumber;
-    }
-    private void moveAndAddToHistory(String direction) {
-        if (!direction.equals("Null") && !Steps.contains(direction)) {
-            Steps.add(currentLocation.getName());
-            currentLocation = getLocation(direction);
-            excludedNumbers.clear();
-        }
-    }
-    private Location getLocation(String loc){
-        for (Location location : mLocations) {
-            String l = location.getName();
-            if (loc.equals(l)){
-                return location;
-            }
-        }
-        return null;
-    }
-
 
 }
