+++ app/src/main/res/layout/navigation.xml
@@ -14,7 +14,7 @@
         android:text="Button"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView3" />
+        app:layout_constraintTop_toBottomOf="@+id/textView3_nav" />
 
     <Button
         android:id="@+id/button2"
@@ -25,12 +25,12 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toEndOf="@+id/button"
-        app:layout_constraintTop_toBottomOf="@+id/textView3" />
+        app:layout_constraintTop_toBottomOf="@+id/textView3_nav" />
 
     <TextView
-        android:id="@+id/textView3"
+        android:id="@+id/textView3_nav"
         android:layout_width="370dp"
-        android:layout_height="89dp"
+        android:layout_height="230dp"
         android:layout_marginTop="25dp"
         android:text="TextView"
         app:layout_constraintEnd_toEndOf="parent"
